Docs: URLs have changed after migration
With hrev51749 the documentation of the package management was
migrated from the Trac wiki into the Haiku tree. The wiki pages
were moved to under 'Obsolete' without redirection.

Removed link to https://dev.haiku-os.org/wiki/PackageManagement/HaikuDepotServer).
That, too was moved to 'Obsolete', but wasn't put into the tree.